Add Pattern
Going forward, having some granularity regarding patterns
(e.g. so that anchored regexes and fixed strings could be mixed)
will be good to have.

More importantly it aids in a separation of responsibilities,
reducing potential code duplication in the constructors for Matcher
(which needs to be made much more flexible anyway).